+++ app/src/main/java/com/kuhrusty/morbadscorepad/SkillListActivity.java
@@ -1,5 +1,6 @@
 package com.kuhrusty.morbadscorepad;
 
+import android.content.Context;
 import android.preference.PreferenceManager;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
@@ -23,6 +24,11 @@ import com.kuhrusty.morbadscorepad.model.Skill;
 import com.kuhrusty.morbadscorepad.model.dao.GameRepository;
 import com.kuhrusty.morbadscorepad.model.dao.RepositoryFactory;
 
+import java.io.BufferedReader;
+import java.io.FileReader;
+import java.io.IOException;
+import java.io.OutputStreamWriter;
+import java.io.Writer;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
@@ -34,7 +40,15 @@ import java.util.List;
 public class SkillListActivity extends AppCompatActivity {
     private static final String LOGBIT = "SkillListActivity";
 
+    private static final String KEY_SELECTED_ADVENTURER = "SkillListActivity.selectedAdventurer";
     private static final String KEY_SELECTED_SKILL = "SkillListActivity.selectedSkill";
+    /**
+     * Because we expect the user to navigate away from this activity (like to
+     * fiddle with the danger deck), we can't always count on
+     * onSaveInstanceState() to store the currently selected adventurer & skill,
+     * so use a file when those guys aren't in the saved instance state.  Ugh.
+     */
+    private static final String STATE_FILENAME = "SkillListActivity.skill.txt";
 
     private GameConfiguration config;
     private GameRepository grepos = RepositoryFactory.getGameRepository();
@@ -48,10 +62,15 @@ public class SkillListActivity extends AppCompatActivity {
     private ImageView masteryIV;
     private String allSkillsItemLabel;
 
+    private String selectedAdventurer;  //  may be null, even when skill is not
+    private String selectedSkillID;  //  may be null
+    private String restoringSkillID;  //  ugh, only used during onCreate()
+
     private class SkillRowSelectionListener implements View.OnClickListener {
         @Override
         public void onClick(View view) {
             String skillID = (String)(view.getTag(R.id.skillID));
+            selectedSkillID = skillID;
             loadImage(skillIV, "skill_" + skillID);
             loadImage(masteryIV, "mastery_" + skillID);
             if (selectedRow != null) selectedRow.setSelected(false);
@@ -66,11 +85,16 @@ public class SkillListActivity extends AppCompatActivity {
         setContentView(R.layout.activity_skill_list);
         allSkillsItemLabel = getString(R.string.skill_list_all_item);
 
-        String selectedSkill = null;
         if (savedInstanceState != null) {
-            selectedSkill = savedInstanceState.getString(KEY_SELECTED_SKILL);
-Log.d(LOGBIT, "onCreate() got savedInstanceState with selected skill " + selectedSkill);
+            selectedAdventurer = savedInstanceState.getString(KEY_SELECTED_ADVENTURER);
+            selectedSkillID = savedInstanceState.getString(KEY_SELECTED_SKILL);
         }
+        if (selectedSkillID == null) {
+            //  can we get these guys from file?
+            loadStateFromFile();
+        }
+        String savedAdventurer = selectedAdventurer;
+        restoringSkillID = selectedSkillID;
 
         if (config == null) {
             config = new GameConfiguration(this, PreferenceManager.getDefaultSharedPreferences(this), grepos);
@@ -120,40 +144,57 @@ Log.d(LOGBIT, "onCreate() got savedInstanceState with selected skill " + selecte
             @Override
             public void onItemSelected(AdapterView<?> adapterView, View view,
                                        int position, long id) {
-                Log.d(LOGBIT, adapterView.getSelectedItem().toString() + " selected");
-                updateSkillList(adapterView.getSelectedItem().toString());
+                String ts = adapterView.getSelectedItem().toString();
+                Log.d(LOGBIT, ts + " selected");
+                selectedAdventurer = ((ts != null) &&
+                        (!ts.equals(allSkillsItemLabel))) ? ts : null;
+                //  Well, this is a ridiculous kludge.  During startup, we want
+                //  to restore the skill they were looking at last time this was
+                //  running.  However, we get *two* selection events during
+                //  startup (even without the explicit
+                //  characterChooser.setSelection(pos); below); the first is
+                //  with view == null, so let's continue to remember the skill
+                //  ID we want to restore during the *second* selection event.
+                //  After that, we'll forget the skill ID we were restoring, so
+                //  that when the *user* clicks on the spinner, we just select
+                //  that adventurer's first skill.
+                selectedSkillID = restoringSkillID;
+                if (view != null) restoringSkillID = null;
+                updateSkillList();
             }
             @Override
             public void onNothingSelected(AdapterView<?> adapterView) {
                 Log.d(LOGBIT, "nothing selected");
             }
         });
-        restoreSelectedSkill(selectedSkill);
+        if (savedAdventurer != null) {
+            selectedAdventurer = savedAdventurer;
+            int pos = adapter.getPosition(selectedAdventurer);
+            if (pos != -1) characterChooser.setSelection(pos);
+        }
+    }
+
+    @Override
+    protected void onStop() {
+        super.onStop();
+        saveStateToFile();
     }
 
     @Override
     public void onSaveInstanceState(Bundle savedInstanceState) {
         super.onSaveInstanceState(savedInstanceState);
 
-        if (selectedRow != null) {
-Log.d(LOGBIT, "onSaveInstanceState() saving selected skill " + selectedRow.getTag(R.id.skillID));
-            savedInstanceState.putString(KEY_SELECTED_SKILL,
-                    (String)(selectedRow.getTag(R.id.skillID)));
+        if (selectedAdventurer != null) {
+            savedInstanceState.putString(KEY_SELECTED_ADVENTURER, selectedAdventurer);
+        }
+        if (selectedSkillID != null) {
+            savedInstanceState.putString(KEY_SELECTED_SKILL, selectedSkillID);
         }
     }
 
-    @Override
-    public void onRestoreInstanceState(Bundle savedInstanceState) {
-        super.onRestoreInstanceState(savedInstanceState);
-
-        restoreSelectedSkill(savedInstanceState.getString(KEY_SELECTED_SKILL));
-    }
-
-    private void updateSkillList(String selectedName) {
-Log.d(LOGBIT, "updateSkillList(" + selectedName + ") hit");
-        AdventurerSheet ta = ((selectedName != null) &&
-                         (!selectedName.equals(allSkillsItemLabel))) ?
-                characterMap.get(selectedName) : null;
+    private void updateSkillList() {
+        AdventurerSheet ta = (selectedAdventurer != null) ?
+                characterMap.get(selectedAdventurer) : null;
         if (ta != null) {
             filteredSkills = Skill.filterSkills(allSkills, ta);
         } else {
@@ -173,23 +214,26 @@ Log.d(LOGBIT, "updateSkillList(" + selectedName + ") hit");
             table.addView(row);
         }
         table.requestLayout();  //  necessary?
-        table.getChildAt(0).callOnClick();
+        if ((selectedSkillID == null) || (!restoreSelectedSkill(selectedSkillID))) {
+            table.getChildAt(0).callOnClick();
+        }
     }
 
-    private void restoreSelectedSkill(String skillID) {
-Log.d(LOGBIT, "restoreSelectedState(" + skillID + ") hit, table == " + table);
+    /**
+     * Returns true if the given skill row was found & selected.
+     */
+    private boolean restoreSelectedSkill(String skillID) {
         if ((skillID == null) || (table == null)) {
-            return;
+            return false;
         }
         for (int ii = 0; ii < table.getChildCount(); ++ii) {
             String tag = (String)(table.getChildAt(ii).getTag(R.id.skillID));
-Log.d(LOGBIT, "  " + tag);
             if ((tag != null) && tag.equals(skillID)) {
-Log.d(LOGBIT, "    CALLING ON CLICK");
                 table.getChildAt(ii).callOnClick();
-                break;
+                return true;
             }
         }
+        return false;
     }
 
     //  This assumes two different skill cards with the same name have the same
@@ -205,5 +249,59 @@ Log.d(LOGBIT, "    CALLING ON CLICK");
     private void loadImage(ImageView iv, String resID) {
         int id = getResources().getIdentifier(resID, "drawable", getPackageName());
         iv.setImageDrawable(getResources().getDrawable((id != 0) ? id : R.drawable.skill_no_image));
+    }
+
+    private void saveStateToFile() {
+        Writer out = null;
+        try {
+            out = new OutputStreamWriter(openFileOutput(STATE_FILENAME, Context.MODE_PRIVATE));
+            out.write("1\t" + selectedAdventurer + "\t" + selectedSkillID + "\n");
+        } catch (IOException ioe) {
+            Log.w(LOGBIT, ioe);
+        } finally {
+            if (out != null) {
+                try {
+                    out.close();
+                } catch (IOException ioe) {
+                    Log.w(LOGBIT, ioe);  //  care *some*, as write may be bad
+                }
+            }
+        }
+    }
+
+    /**
+     * This is just here to keep some clutter out of onCreate().
+     */
+    private void loadStateFromFile() {
+        BufferedReader in = null;
+        String line = null;
+        try {
+            in = new BufferedReader(new FileReader(openFileInput(STATE_FILENAME).getFD()));
+            line = in.readLine();
+        } catch (Exception ex) {
+            //  we don't really care; the first time this runs, we expect
+            //  FileNotFoundException, and after that, if we can't read the
+            //  file, they just don't get their state restored.
+        } finally {
+            if (in != null) {
+                try {
+                    in.close();
+                } catch (IOException ioe) {
+                    Log.w(LOGBIT, ioe);  //  don't care
+                }
+            }
+        }
+        if (line == null) return;
+        //  Real classy: we expect, tab-delimited, a version number, an
+        //  adventurer, and a skill ID.  If we find anything unexpected, bail,
+        //  because it really doesn't matter if their state can't be restored
+        //  this one time.
+        String[] bits = line.split("\t");
+        if (bits.length != 3) return;
+        selectedAdventurer = bits[1];
+        selectedSkillID = bits[2];
+        if ((selectedAdventurer != null) && selectedAdventurer.equals("null")) {
+            selectedAdventurer = null;
+        }
     }
 }